+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/DashInstance.java
@@ -521,15 +521,15 @@ public class DashInstance extends EntityTameable {
     }
 
     public boolean getCanSpawnHere() {
-        for (byte k = -3; k < 3; ++k) {
-            for (byte j = -3; j < 3; ++j) {
-                for (byte i = 0; i < 5; ++i) {
+        for (int k = -3; k < 3; ++k) {
+            for (int j = -3; j < 3; ++j) {
+                for (int i = 0; i < 5; ++i) {
                     final Block bid = this.worldObj
-                        .getBlock((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                        .getBlock((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                     if (bid == Blocks.mob_spawner) {
                         TileEntityMobSpawner tileentitymobspawner;
                         tileentitymobspawner = (TileEntityMobSpawner) this.worldObj
-                            .getTileEntity((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                            .getTileEntity((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                         final String s = tileentitymobspawner.func_145881_a()
                             .getEntityNameToSpawn();
                         if (s != null && s.equals("MyDash")) {
@@ -1132,7 +1132,7 @@ public class DashInstance extends EntityTameable {
                     p.getUniqueID()
                         .toString());
                 this.playTameEffect(true);
-                this.worldObj.setEntityState((Entity) this, (byte) 7);
+                this.worldObj.setEntityState((Entity) this,  (byte) 7);
                 this.heal(this.mygetMaxHealth() - this.getHealth());
             }
             if (!par1EntityPlayer.capabilities.isCreativeMode) {
@@ -1157,11 +1157,11 @@ public class DashInstance extends EntityTameable {
                         p.getUniqueID()
                             .toString());
                     this.playTameEffect(true);
-                    this.worldObj.setEntityState((Entity) this, (byte) 7);
+                    this.worldObj.setEntityState((Entity) this,   (byte)7);
                     this.heal(this.mygetMaxHealth() - this.getHealth());
                 } else {
                     this.playTameEffect(false);
-                    this.worldObj.setEntityState((Entity) this, (byte) 6);
+                    this.worldObj.setEntityState((Entity) this,   (byte)6);
                 }
             }
             if (!par1EntityPlayer.capabilities.isCreativeMode) {
@@ -1185,7 +1185,7 @@ public class DashInstance extends EntityTameable {
         if (var2 != null && par1EntityPlayer.getDistanceSqToEntity((Entity) this) < 49.0) {
             if (this.worldObj.isRemote) {
                 this.playTameEffect(true);
-                this.worldObj.setEntityState((Entity) this, (byte) 7);
+                this.worldObj.setEntityState((Entity) this, (byte)  7);
             }
             if (this.mygetMaxHealth() > this.getHealth()) {
                 this.heal(this.mygetMaxHealth() - this.getHealth());
@@ -1205,7 +1205,7 @@ public class DashInstance extends EntityTameable {
                 this.setTamed(false);
                 this.func_152115_b("");
                 this.playTameEffect(false);
-                this.worldObj.setEntityState(this, (byte) 6);
+                this.worldObj.setEntityState(this,  (byte) 6);
             }
             if (!par1EntityPlayer.capabilities.isCreativeMode) {
                 final ItemStack itemStack4 = var2;
