+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/PhoenixInstance.java
@@ -496,15 +496,15 @@ public class PhoenixInstance extends EntityTameable {
     }
 
     public boolean getCanSpawnHere() {
-        for (byte k = -2; k <= 2; ++k) {
-            for (byte j = -2; j <= 2; ++j) {
-                for (byte i = 1; i < 4; ++i) {
+        for (int k = -2; k <= 2; ++k) {
+            for (int j = -2; j <= 2; ++j) {
+                for (int i = 1; i < 4; ++i) {
                     final Block bid = this.worldObj
-                        .getBlock((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                        .getBlock((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                     if (bid == Blocks.mob_spawner) {
                         TileEntityMobSpawner tileentitymobspawner;
                         tileentitymobspawner = (TileEntityMobSpawner) this.worldObj
-                            .getTileEntity((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                            .getTileEntity((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                         final String s = tileentitymobspawner.func_145881_a()
                             .getEntityNameToSpawn();
                         if (s != null && s.equals("MyPhoenix")) {
@@ -1223,7 +1223,7 @@ public class PhoenixInstance extends EntityTameable {
                     par1EntityPlayer.getUniqueID()
                         .toString());
                 this.playTameEffect(true);
-                this.worldObj.setEntityState((Entity) this, (byte) 7);
+                this.worldObj.setEntityState((Entity) this,  (byte) 7);
                 this.heal(this.getMaxHealth());
                 this.heal(this.mygetMaxHealth() - this.getHealth());
             }
@@ -1247,11 +1247,11 @@ public class PhoenixInstance extends EntityTameable {
                             par1EntityPlayer.getUniqueID()
                                 .toString());
                         this.playTameEffect(true);
-                        this.worldObj.setEntityState((Entity) this, (byte) 7);
+                        this.worldObj.setEntityState((Entity) this,  (byte) 7);
                         this.heal(this.mygetMaxHealth() - this.getHealth());
                     } else {
                         this.playTameEffect(false);
-                        this.worldObj.setEntityState((Entity) this, (byte) 6);
+                        this.worldObj.setEntityState((Entity) this,  (byte) 6);
                     }
                 }
                 if (!par1EntityPlayer.capabilities.isCreativeMode) {
@@ -1280,7 +1280,7 @@ public class PhoenixInstance extends EntityTameable {
                 && par1EntityPlayer.getDistanceSqToEntity((Entity) this) < 49.0) {
                 if (this.worldObj.isRemote) {
                     this.playTameEffect(true);
-                    this.worldObj.setEntityState((Entity) this, (byte) 7);
+                    this.worldObj.setEntityState((Entity) this,   (byte)7);
                 }
                 if (this.mygetMaxHealth() > this.getHealth()) {
                     this.heal(this.mygetMaxHealth() - this.getHealth());
@@ -1301,7 +1301,7 @@ public class PhoenixInstance extends EntityTameable {
                     this.setTamed(false);
                     this.func_152115_b("");
                     this.playTameEffect(false);
-                    this.worldObj.setEntityState((Entity) this, (byte) 6);
+                    this.worldObj.setEntityState((Entity) this,  (byte) 6);
                 }
                 if (!par1EntityPlayer.capabilities.isCreativeMode) {
                     final ItemStack itemStack4 = var2;
