+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/Rat.java
@@ -310,15 +310,15 @@ public class Rat extends EntityMob {
 
     public boolean getCanSpawnHere() {
         int sc = 0;
-        for (byte k = -2; k < 2; ++k) {
-            for (byte j = -2; j < 2; ++j) {
-                for (byte i = 0; i < 5; ++i) {
+        for (int k = -2; k < 2; ++k) {
+            for (int j = -2; j < 2; ++j) {
+                for (int i = 0; i < 5; ++i) {
                     final Block bid = this.worldObj
-                        .getBlock((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                        .getBlock((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                     if (bid == Blocks.mob_spawner) {
                         TileEntityMobSpawner tileentitymobspawner;
                         tileentitymobspawner = (TileEntityMobSpawner) this.worldObj
-                            .getTileEntity((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
+                            .getTileEntity((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
                         final String s = tileentitymobspawner.func_145881_a()
                             .getEntityNameToSpawn();
                         if (s != null && s.equals("Rat")) {
